Rename SelectorSourceTransaction to CandidateTransaction and move it under MiningManager
