+++ mining/src/mempool/model/transactions_pool.rs
@@ -1,11 +1,10 @@
 use crate::{
-    block_template::selector::SelectorSourceTransaction,
     mempool::{
         config::Config,
         errors::{RuleError, RuleResult},
         model::{map::MempoolTransactionCollection, pool::Pool, tx::MempoolTransaction, utxo_set::MempoolUtxoSet},
     },
-    model::topological_index::TopologicalIndex,
+    model::{candidate_tx::CandidateTransaction, topological_index::TopologicalIndex},
 };
 use consensus_core::{
     api::DynConsensus,
@@ -212,13 +211,11 @@ impl TransactionsPool {
 
     /// all_ready_transactions returns all fully populated mempool transactions having no parents in the mempool.
     /// These transactions are ready for being inserted in a block template.
-    pub(crate) fn all_ready_transactions(&self) -> Vec<SelectorSourceTransaction> {
+    pub(crate) fn all_ready_transactions(&self) -> Vec<CandidateTransaction> {
         // The returned transactions are leaving the mempool so they are cloned
         self.all_transactions
             .values()
-            .filter_map(
-                |x| if self.is_transaction_ready(&x.id()) { Some(SelectorSourceTransaction::from_mutable(&x.mtx)) } else { None },
-            )
+            .filter_map(|x| if self.is_transaction_ready(&x.id()) { Some(CandidateTransaction::from_mutable(&x.mtx)) } else { None })
             .collect()
     }
