+++ mining/src/model/mod.rs
@@ -1,6 +1,7 @@
 use consensus_core::tx::TransactionId;
 use std::collections::HashSet;
 
+pub(crate) mod candidate_tx;
 pub mod owner_txs;
 pub mod topological_index;
